+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './contentContainer/contentContainer.component';
 export * from './grid/grid.component';
 export * from './gridContainer/gridContainer.component';
 export * from './matrixGrid/matrixGrid.component';