+++ src/directives/index.ts
@@ -1,3 +1,4 @@
+export * from './contentContainerTemplate/contentContainerTemplate.directive';
 export * from './gridContainerTemplate/gridContainerTemplate.directive';
 export * from './orderable/orderable.directive';
 export * from './tableGridBodyCellTemplate/tableGridBodyCellTemplate.directive';